# Optimize cron parsing