## Quartz.Benchmark.CronExpressionBenchmark

| **Diff**|Method|Toolchain|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |CreateExpressionsAndCalculateFireTimeAfter|Default|12.56 μs|0.129 μs|21.56 KB|
| **New** |	| **Default** | **10.687 μs (-15%)** | **0.1867 μs** | **16.48 KB (-24%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|Default|12.55 μs|0.047 μs|24.56 KB|
| **New** |	| **Default** | **10.354 μs (-17%)** | **0.1038 μs** | **18.28 KB (-26%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|Default|11.93 μs|0.148 μs|19.02 KB|
| **New** |	| **Default** | **9.681 μs (-19%)** | **0.1570 μs** | **15.04 KB (-21%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|Default|13.13 μs|0.074 μs|25.78 KB|
| **New** |	| **Default** | **10.763 μs (-18%)** | **0.0749 μs** | **19.23 KB (-25%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|Default|14.83 μs|0.208 μs|26.07 KB|
| **New** |	| **Default** | **10.439 μs (-30%)** | **0.0860 μs** | **19.34 KB (-26%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|Default|13.26 μs|0.135 μs|26.07 KB|
| **New** |	| **Default** | **11.291 μs (-15%)** | **0.1250 μs** | **19.34 KB (-26%)** |